New Development: No             JIRA Issue: Yes CAS-4827         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
TableRecord objects could not be used with the templated ScalarColumnData<T>
object since the needed operator== method is not present.  However, since
== is only called if a sentinel "undefined" value option is provided
in the ScalarColumnData constructor, it will work to provide an == operator
which throws if called (the effort required to build a correct == is not
warranted at this time).