New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Fleshed out the VisBuffer2Adapter class so that it provides adequate
functionality for use when on-the-fly model generation is used.  Since
this was a kluge to allow use of the existing model generation code
which accepts VisBuffer(1) objects, only enough VB functionality was
implemented to get things to work (and they appear to do so now).