New Development: No             JIRA Issue: CAS-6198         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: test_importasdm   Put in Release Notes: No              Module(s): alma            Description: two important changes.
the data produced were re ordered by increasing values of data description index. We did not want that, we wanted to have them ordered as they are ordered in the ASDM ConfigDescription table. We have modified the code so that the ordering is preserved.

 
the calculations of the cross correlation values are now done exactly as they are done by the lazz filler (the formulas are stricty the same). This was a required condition to validate the comparison of the fillers.