New Development: Yes             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): CASA.cmake in build system            Description:
The macro "qt4_extract_options" has changed signature by adding a
"target" argument.  This macro is used by casa_qt4_wrap_ui.  The
result is that the first file provided to the casa_qt4_wrap_ui macro
is lost and during the "make" the first file's .ui.h file is not
generated which shows up as a missing include during c++ compilation.
This mod into our CASA.cmake file uses the new calling
qt4_extract_options macro if the cmake version is > 2.8.5 and
otherwise uses the original one.  It's possible that the cmake version
needed to trigger the old vs. new qt4_extract_version signature change
is higher than 2.8.5 (but less than 2.8.12.2); tracking it down
doesn't seem to be straightfoward but maybe that's just due to my
ignorance of git; in any case, we'll discover this one the hard way
(if we're unlucky).