New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): MSVis            Description:
Modified VI/VB2 framework to better support TransformingVi2 which is
being used by clibrating VI effort.  This should make it easier for the
descendants of transforming VI2 to properly configure their associated
VisBuffer2s via the new configureNewSubchunk methods added.

I'm in the middle of other work in the MSVis area, so I had to pick and
choose which modified files to commit.  It's possible in doing so I will
break the build even though my own is fine; only one way to find out, 
though ;-(