New Development: Yes             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s):            Description:
Modified VI2 framework to improve the API for accessing the
correlation types.  Two new methods were added to VisBuffer2:
getCorrelationTypesDefined and getCorrelationTypesSelected; both
return a Vector<Stokes::StokesTypes>.  The first method returns all
correlation types defined for the current spectral window while the
second returns the types that are actually selected (in the order that
they appear in the various cubes).  The older methods correlationTypes
and getCorrelation types will be removed in a future release.