New Development: Yes             JIRA Issue: CAS-7623         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: test_importasdm   Put in Release Notes: No              Module(s):            Description: Added support for spw selection in the online flags created by importasdm. If Flag.xml contains numSpectralWindow > 0, it will create spw syntax by name instead of index. The parseXML will skip the WVR#Antenna_* if WVR#Nominal is present in the current flag.
Together with this, I'm removing the calculation of a union of the data selection parameters when flagdata is called in list mode. With the addition of the spw selection above, the call to ms.msseltoindex inside flaghelper::compressSelectionList() gets too slow without any benefit. The compression of flags done by _merge_timerange() aleady does all the improvement in a long list. If side-effects are seen later, the union calculation should be added back.