New Development: No             JIRA Issue: No         Ready for Test:      Interface Changes: What Interface Changed:          Test Programs:   Put in Release Notes:              Module(s):            Description:
Modified build macros to support casa_add_demo which should be used to
define demo applications (these are not unit tests but typically
located with them).  Modified the various module CMakeLists.txt to use
the new macro plus mark a couple of the demo apps as unbuildable.
Made minor changes to a couple demo apps so that they would build.
Looks like the warning count may increase for some modules due to
warnings in the demos.