New Development: No             JIRA Issue: CAS-8204         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: manual tests   Put in Release Notes: Yes/No              Module(s): stdcasa            Description: Modified method casaMDirection() in CasacSupport.cc. This method is used in imager::defineimage() to convert the phasecenter string to an MDirection. If the Dec/Lat part of the phasecenter string is on the one hand correctly interpreted as an angle but on the other hand contains the character ":" (colon), the user now receives a notification
  NOTE: you provided the Declination/Latitude value "14:03:23.123" which is understood to be in units of "hours", not degrees!

and the program proceeds.
This will not break any code (I checked: casaMDirection() is not used in many places)
and will clarify to the user that the colon syntax implies "hours", not "degrees".
If the user enters 14h03m23.123 for the Declination, there will be no warning as the user has explicitly stated the units.