New Development: No             JIRA Issue: No         Ready for Test:      Interface Changes: What Interface Changed:          Test Programs:   Put in Release Notes:              Module(s):            Description:
Suppressed warnings in the bnmin1 module.  Note that a couple routines
have signatures indicating they returned "double" but had no return
statement and did not have anything obvious to return: added a return
statement that returns a signalling NaN (this ought to generate an
exception in any code that is actually using the return value which
will trigger further investigation).