New Development: Yes             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes:              Module(s):            Description:
Put in support to build the crash reporter functionality.  These mods will
build normally (i.e., without the crash reporter logic) unless the
argument -DUseCrashReporter=1 is provided during cmake to both the code
and gcwrap projects.

When UseCrashReporter=1 is provided code will download the breakpad tar
file from a provided CASA server location into a directory in the code
build tree.  It will then build breakpad and provide cmake symbols
allowing other projects to find the breakpad static library and its
include files.

To install the crashreporter logic into the casapy executable the same
cmake argument needs to be provided there.  If it is provided in gcwrap
then there will be build errors if it wasn't provided to the code build.
When installed the crash reporter logic will enable crash reporting in
casapy if it dies from a segfault or unhandled exception, etc.

Once I am confident it will build and test on all of our supported systems
and the needed web infrastructure is in place, the default value of the
UseCrashReporter logic will be flipped and it will be built in by default.
At this commit there has been no Apple testing, so the functionality needs
to be disabled by default.