New Development: No             JIRA Issue: http://jira.alma.cl/browse/PRTSIR-9678         Ready for Test: Yes/No      Interface Changes: Yes/No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: Describe your changes here...
ASDM and MS use to store the baselines data for one integration in transposed orders. This convention is respected by the application asdm2MS (filler) which performs the transposition on the fly as it fills the MS Main table. But bdflags2MS was forgetting to perform this transposition while it distributed the flags found in the ASDM BDFs over the MS Main rows of column FLAG. This version corrects this bug.