CASA3K: add standard __future__ imports for site-packages gcwrap files
Namely, `from __future__ import absolute_import, division, print_function`.

When packaging on Python 2.7, there are complaints because Conda tries to
byte-compile some of the files installed in scripts/casa-python/tasks, which
do not have these imports, leading to problems with print-as-a-function. But
these are nonfatal, and the files in that directory are intended to be
provided as guides only. So I'm ignoring them for now.

(This addition was done manually; if new releases churn things a lot, it might
help to figure out how to automate it.)