Preliminary implementation up to pre-application
CAS-8879
Preliminaty implementation of SDDoubleCircleGainCal that is 
compatible with new Calibrater. Up to pre-application has 
implemented. selfSolveOne is written so that it only copies 
given corrected visibility into calibration solution. 
Note that class's behavior changes depending on USEOLDVI 
macro in the header. To pass sdgaincal unit test, USEOLDVI 
must be true. On the other hand, new implementation requires 
to set USEOLDVI to false.