Add a test suite for asdmsummary:  CAS-9883
This is a trivial test suite for the asdmsummary task, which summarizes an
ASDM to the logger.  The tests are done on several ASDM.  The tests also
check that the expected number of new rows were written to the logger.
No checks are done on the content of the logged lines.