A draft of the fringe fitter.
Currently it uses the GSL Levenberg-Marquardt algorithm, but doesn't use weights.
It solves for phase at the min(t), min(f) corner of the grid and shifts the result
to the reference values as a last step.

It seems to work on my one simple test-case; it is not at all finished.