Changed handling of baseline sign in FFT stage, and bug fixes.
I was taking conjugants of complex numbers to handle the baselines where the reference
antenna came second. I am now changing the sign of the results instead.

Many other bug fixes, notably I was not including the reference antenna in the list of
active antennas, but I was subtracting one from the number of antennas to form the
number of parameters.