Move VisBuffer2 creation method to ViImplementation2
ViImplementation2 instances should know how to create the correct VisBuffer2 for
themselves, so remove the VisBufferType enumeration, and implement a non-pure,
virtual ViImplementation2::createAttachedVisBuffer() method for that
purpose. The two argument VisBuffer2::factory() method is maintained to allow
for creating VisBuffer2 instances that are not attached to a ViImplementation2
instance.