Added multiband fringe fitting and "manual phase cal" support.
Multiband fringe fitting is implemented using the built-in "spw" option to
standard calibrater "combine" keyword; manual phase calibration is implemented
by setting delay-rates to zero after solving. It requires a new boolean
parameter 'zerorates' to be set to True for it to be turned on. Adding this
new parameter involved changing Calibrater.cc and others to get it threaded
through the calibration architecture.