Add TryBase class
To be used by Try template methods to test whether Try::value_class is a Try
type. This allows better template parameter deduction, and allows the
Try::flatten method to be called without an explicit parameter in most
cases. With this change, the Try::flatMap and Try::transform methods have been
revised to call the flatten method.