Reorganize polflag test
CAS-10683
Expected result for test_i has been changed for new image framework that
flags all correlations when any of them are flagged so that it produces
empty image when one of correlations are fully flagged. 

Defined test_pseudo_i for pseudo-Stokes mode. This case is consistent
with older imager with stokes='I'. At this moment, the test is skipped
since pseudo-Stokes is not implemented yet.