Performance improvement of sdcal apply mode
CAS-11056
By using Intel VTune Amplifier, I detected bottleneck. Slowness of apply
operation was due to too frequent caltable selection. I defined a cache
for necessary information in caltable and keep it as a list of CASA
Arrays that are connected to a set of IDs (observation ID, spw ID, and
antenna ID). By using cache, caltable selection hppens only once for
each combinations of IDs. As a result, most of the CPU time were spent
for apply operation and I/O which is reasonable.