Removing description relevant to ASAP
CAS-11114
There is a description about ATNF Spectral Analysis Package, ASAP, in an
inline help for sdimaging and tsdimaging. The sentence just says that
these tasks are independent of ASAP. Nowadays, however, all the single
dish tasks are based on MS and CASA no longer depends on ASAP. Thus,
that part must be removed.