Do not consider the antenna selection trivial if output antennas are fewer than input
There is some logic to check whether the antenna selection
is trivial, i.e., there has been some baseline selection
but the antennas used are the same as in the input.
The logic was flawed in the case the number of total antennas
was less than the input.
Fixes partially CAS-11111