Support setting of frequency selections in the factory
The reasoning behind this is the following: suppose a scenario
where we have an underlying MS with a channel selection and
a TVI stack that operates on that. Before thie patch you could
apply channel selection to the generated VI, but that's the
top VI, the one that operates on the already transformed
data by the other underlying TVI. You want the channel selections
to operate on the first layer that access the MS directly. The
rest of the layers just see what the layer below provide and
they don't need channel selection.
This commit allows the "disk layer" VI to apply those selections
to the MS.
Fixes partially CAS-9679