Do not parse the spw parameter from the configuration
This commit is the meat of CAS-9679. It no longer parses the spw parameter.
Originally this was used to pass the MS selection on spw and channels.
However that breaks the layering principle. Now, it simply asks the previous
layer for the information of which SPWs and channels are available.
Fixes partially CAS-9679