Make the configuration parameter of the ConvolutionTVI base class optional
Once the spw parameter has been removed, there is no need for some derived
TVIs, like HanningSmooth to even have such parameter, let alon passing it
to the base class. Actually it can even be a problem if a user specifies
the "kernel" parameter in the configuration and this is passed by the
derived class to the base class without sanitizing the user input.
Fixes partially CAS-9679