Throw an exception also if the data cube has not been computed
The time averaging TVI has the option to selectively average
some data cubes only. If a data cube that has not been averaged
is requested there would be a non-consistent view from the upper
TVI layer in which some cubes are time averaged and others not.
This commit mitigates that by throwing an exception is a cube
that has not been computed is requested. Note that these cubes are still
available if using the VisBuffer directly.
Fixes partially CAS-8222