Use the new methods in the disk layer VI to set the channel selection.
As per commit 01242d1c6d789cb4502dd32da3d224191bd1b65e
the channel selection should be applied to the first layer that
operates on the MS directly, which is the disk IO layer. If the
channel selection is applied to the top level VI, then the selection
happens on the potentially modified SPWs, which probably not the
intention.

Fixes partially CAS-9679