Workaround issue with CAS-11734: getChannels() throws if phantom SPWs are present
CAS-11734 is causing some problems with the FreqAxisTVI class in the
presence of SPWs which are not mentioned in the DDI table.
The workaround here it to filter the spws before calling gaincal()
to avoid the exception. Since those phantom SPWs are not used by the
subsequent analysis this should be fine.

Fixes partially CAS-9679