Implement the creation of subtables antenna, spw, data description and polarization.
This commit implements the overriding of the subtables accesors for
this simulation VI. Note that all except the four mentioned
subtables throw an exception. For the four mentiond subtables a
very simplistic table is returned, with just the minimum consistency,
but without any real useful value for most of the columnns.

Fixes partially CAS-11992