Deprecate spwOutChanNumMap_p and use more explicit spwInpChanIdxMap_p
The spwOutChanNumMap_p map contained the number of channels per
output SPW. The new spwInpChanIdxMap_p is more explicit, in the sense
that it also contains the list of channel indices. The number of channels
can still be accessed using the .size() method of the indices vector.

Fixes partially CAS-11988