Remove mutable keyword
It is not recommended to use the mutable keyword unless there
are good justifications for it. In this case it was even dangerous,
since std::map would silently create a new entry if the lookup of the key
fails. This change has actually revealed that this was the case
in the unit tests.

Fixes partially CAS-11988