Fix the unit tests after changing getChannels() from FreqAxisTVI
The propagation of the SPW subtable has the side effect of
changing the output channel map, for which getChannels() had to be changed.
Additionally nSpectralWindows() in ChannelAverageTVI returns more SPWs
than before (basically the double, since the old ones are kept).

Those changes have to be reflected in the unit test expectations.

Fixes CAS-11988