gcwrap: fix Python modules on macOS/py37
The problem is that now, if a module links with libpython, you will get a
segfault. This seems to be the case not only for the various CASA Python
extension modules themselves but also for CASA's `libtools.dylib`, which they
link to. I've copied the linker flags manually since I can't figure out how to
figure them out on-the-fly in any reasonable way.