This is another large commit with the intention of converting the pointing offset object from a vector of doubles to a vector of vector of doubles. This required template changes to the following
 code/synthesis/TransformMachines2/ConvolutionFunction.h
 code/synthesis/TransformMachines2/IlluminationConvFunc.h
 code/synthesis/TransformMachines2/AWConvFunc.h
 code/synthesis/TransformMachines2/CFBuffer.h

Changes were also made to code/synthesis/TransformMachines2/CFBuffer.cc
in the class toAsStruct a GPU convolution function generation class.

The remaining changes to code/synthesis/TransformMachines2/AWProjectFT.cc
 code/synthesis/TransformMachines2/AWProjectWBFT.cc was to ensure the
PointingOffset object was piped into the mapping classes.

Changes in PointingOffsets PhaseGrad and VB2CFBMap classes are for the purpose
of implementing the vector of vectors pointing object along with indexing to
ensure the return data types are matched.

The regression for this code is to follow shortly after