CAS-12391 CASA 5 does not honor default vector values in
eg cs.newcoordsys() stokes, and anyway this parameter should be
empty by default, so its not clear to me why it had values in
the first place. CASA 6 honors these values and now that they
are appearing, it is breaking tests. By default, stokes in this
method should be empty.